Makes MetadataAwareValue an inner class of EntryLoader
