+++ hazelcast/src/main/java/com/hazelcast/map/impl/mapstore/writebehind/entry/NullValueDelayedEntry.java
@@ -16,7 +16,7 @@
 
 package com.hazelcast.map.impl.mapstore.writebehind.entry;
 
-import com.hazelcast.map.MetadataAwareValue;
+import com.hazelcast.map.EntryLoader;
 import com.hazelcast.map.impl.mapstore.writebehind.WriteBehindStore;
 
 /**
@@ -48,7 +48,7 @@ class NullValueDelayedEntry<K, V> implements DelayedEntry<K, V> {
 
     @Override
     public long getExpirationTime() {
-        return MetadataAwareValue.NO_TIME_SET;
+        return EntryLoader.MetadataAwareValue.NO_TIME_SET;
     }
 
     @Override
