+++ hazelcast/src/test/java/com/hazelcast/map/ExtendedValueTest.java
@@ -16,6 +16,7 @@
 
 package com.hazelcast.map;
 
+import com.hazelcast.map.EntryLoader.MetadataAwareValue;
 import com.hazelcast.test.HazelcastParallelClassRunner;
 import com.hazelcast.test.annotation.ParallelJVMTest;
 import com.hazelcast.test.annotation.QuickTest;
@@ -23,7 +24,7 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 
-import static com.hazelcast.map.MetadataAwareValue.NO_TIME_SET;
+import static com.hazelcast.map.EntryLoader.MetadataAwareValue.NO_TIME_SET;
 import static org.junit.Assert.assertEquals;
 
 @RunWith(HazelcastParallelClassRunner.class)
