+++ hazelcast/src/test/java/com/hazelcast/map/impl/mapstore/writebehind/EntryStoreWriteBehindCustomClockTest.java
@@ -19,7 +19,7 @@ package com.hazelcast.map.impl.mapstore.writebehind;
 import com.hazelcast.config.Config;
 import com.hazelcast.config.MapStoreConfig;
 import com.hazelcast.core.HazelcastInstance;
-import com.hazelcast.core.IMap;
+import com.hazelcast.map.IMap;
 import com.hazelcast.map.impl.mapstore.TestEntryStore;
 import com.hazelcast.test.HazelcastSerialClassRunner;
 import com.hazelcast.test.annotation.QuickTest;
